pyproject.toml
[tool.poetry]
name = "pydata_tokyo_24"
version = "0.1.0"
description = ""
authors = ["Yasunori Horikoshi <hotoku@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = ">=3.9, <3.11"
bqsc = "^3.1.0"
pandas-gbq = "^0.16.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
